// backend/src/enums/revenuecat-event.enum.ts
export enum RevenueCatWebhookEvent {
   INITIAL_PURCHASE = "INITIAL_PURCHASE",
   RENEWAL = "RENEWAL",
   PRODUCT_CHANGE = "PRODUCT_CHANGE",
   CANCELLATION = "CANCELLATION",
   BILLING_ISSUE = "BILLING_ISSUE",
   UNCANCELLATION = "UNCANCELLATION",
   EXPIRATION = "EXPIRATION",
   REFUND = "REFUND"
}